backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
feedparser==6.0.10
python-multipart==0.0.6
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai>=0.3.0
